+++ Source/YaOpt/Helpers/ConcurrentPool.cs
@@ -2,20 +2,20 @@ using System.Collections.Concurrent;
 
 namespace YaOpt.Helpers
 {
-	internal static class ConcurrentPool<T> where T : new()
+	public static class ConcurrentPool<T> where T : new()
 	{
-		public static ConcurrentBag<T> Bag = new ConcurrentBag<T>();
+		private static readonly ConcurrentBag<T> _bag = new ConcurrentBag<T>();
 
-		public static int FreeItemsCount => Bag.Count;
+		public static int FreeItemsCount => _bag.Count;
 
 		public static T Get()
 		{
-			return Bag.TryTake(out var result) ? result : new T();
+			return _bag.TryTake(out var result) ? result : new T();
 		}
 
 		public static void Return(T item)
 		{
-			Bag.Add(item);
+			_bag.Add(item);
 		}
 	}
 }